+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 - rendering the task-message-above and task-message-below UI extension placements in compact display mode
 - added icons to archive/unarchive and delete buttons in task actions panel
 - correctly placing unmaximized app window on startup
+- added support for defining supported OS in the extension and showing an info if the extension is not supported on the current OS
+- added support for using external libraries for extension UI components
+- added support for new UI extension placements ('tasks-sidebar-actions-left', 'tasks-sidebar-actions-right') to be used in Tasks header
 
 ## [0.67.0]